requirements.txt
tensorflow
keras
flask
wheel
pandas
pillow
numpy